+++ src/pages/LoadUnload/components/MapLayers/LoadUnloadSpacesLayer.tsx
@@ -1,6 +1,7 @@
 import { useMapInstance } from '@amsterdam/arm-core'
 import { NonTiledLayer } from '@amsterdam/arm-nontiled'
-import { useEffect } from 'react'
+import { LeafletMouseEvent } from 'leaflet'
+import { useCallback, useEffect } from 'react'
 
 import { loadUnloadSpaces } from '../../../../shared/map/mapLayers'
 
@@ -10,17 +11,25 @@ import { useLoadUnloadMapContext } from '../../contexts/MapContext'
 export const LoadUnloadLoadUnloadSpacesLayer = () => {
   const mapInstance = useMapInstance()
   const { activeMapLayers, setDetailFeature } = useLoadUnloadMapContext()
-
-  useEffect(() => {
-    mapInstance.on('click', e => {
+  const onClick = useCallback(
+    (e: LeafletMouseEvent) => {
       if (mapInstance.getZoom() >= loadUnloadSpaces.options.minZoom!) {
         setDetailFeature({
           type: DetailFeatureActionType.SET_LOCATION,
           location: [e.latlng.lat, e.latlng.lng],
         })
       }
-    })
-  }, [mapInstance, setDetailFeature])
+    },
+    [mapInstance, setDetailFeature]
+  )
+
+  useEffect(() => {
+    if (activeMapLayers[loadUnloadSpaces.id]) {
+      mapInstance.on('click', onClick)
+    } else {
+      mapInstance.off('click', onClick)
+    }
+  }, [activeMapLayers, mapInstance, setDetailFeature, onClick])
 
   if (!activeMapLayers[loadUnloadSpaces.id]) return null