+++ CHANGELOG.md
@@ -8,9 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 - Load-Unload: address search
+- Load-Unload: road sections with load-unload information
 
 ### Changed
 - Load-Unload: increase max zoom level to 21
+
+### Fixed
+- Prohibitory signs: prevent undefined maxAllowedWeight for vehicle
 
 
 ## [v2.2.1 - 2022-12-15](https://github.com/Amsterdam/bereikbaarheid-frontend/compare/v2.2...v2.2.1)